+++ doc/release-notes.md
@@ -1,4 +1,5 @@
 # 1.0.2
+- anonymous files handling enabled
 
 # 1.0.1
 - refactoring and unit tests